Create generic database schema tests
The old foreign key constraint test was testing the SQL driver in the prepopulator, not our SQL database schema. Move the foreign key pragma to the SQL runtime code to enforce foreign key constraints on all clients. Also add an overarching test to catch any missing 'IF NOT EXISTS' statements.