Abstract read-only database configuration to SQLiteReadOnlyOpenHelper.
Open the database for reading only, the only reliable way to open a database if the disk may be full.
Provide the readable database when the writable database is requested, as `SQLDelight` only supports writable databases.
https://github.com/cashapp/sqldelight/issues/2163

Don't explicitly turn off database journaling, the journal file is not opened for operations that don't alter the database.
Don't explicitly turn off sync mode, it isn't used in read-only mode.

Upgrade SQLiteCopyOpenHelper to 0.1.1